requirements.txt
streamlit
ffmpeg-python
demucs
torch
